requirements.txt
# --- 数据抓取与解析 (Fetcher) ---
arxiv>=2.1.0        # 用于调用 ArXiv 官方 API 获取论文元数据
requests>=2.31.0    # 用于下载 PDF 原文文件
PyMuPDF>=1.23.0     # 用于极速解析 PDF 提取纯文本 (代码中 import fitz)

# --- 大模型与数据校验 (Reviewer & LLM API) ---
openai>=1.14.0      # 你的 src/llmapi/model.py 网关底层的请求 SDK
pydantic>=2.6.0     # 用于定义结构化的打分模型 (PaperScore, BatchScoringResult)

# --- 配置与工具 (Config & Notifier) ---
python-dotenv>=1.0.1 # 用于安全加载 .env 文件中的账号密码等环境变量
Jinja2>=3.1.3        # 用于渲染精美的 HTML 邮件模板
